Tests/LibWeb/Text/expected/display_list/caret-in-hidden-inline-context.txt

in-block: 1
in-inline: 1
